Closes #2
Issue #2 reported the absence of probabilistic features (such as those
described in Gillick 2009 or Riley 1989). These have been implemented
and incorporated into the program.

It is not obvious to me, however, that they help much. However, I
hypothesize they will be most useful in one particular case: when
the text to be segmented is single-case, but the training data is mixed
case. More on that later.

This commit closes #2.