weight culling
When running from command line (i.e., __main__), null classifer weights
are culled (i.e., the weights are refreshed and any weights which have
an average value of zero are removed). This greatly shrinks the size of
the pickled model.

You have to do this explicitly when using the Python API.