Experimental branch using Perceptronix.
This is the first commit of a new branch which uses Perceptronix (a
C++14 port of my averaged perceptron classifier).

Models are now serialized as protobufs. Deserialization therefore
becomes much faster. The serialization step also culls zero weights, so
the final model is much smaller (60k vs. 500k).

Startup time (i.e., loading the SO) is still pretty good.

Training is a bit slower, but prediction is faster.

This branch also depends on a new (not quite yet released) version of
nlup (0.7.0) which cleans up some bits and pieces there.